pyproject.toml
[project]
name = "auth-test-task"
version = "0.1.0"
description = "Test task"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "email-validator>=2.0.0",
    "fastapi[standard]>=0.135.1",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.12.1",
    "redis>=7.3.0",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]
